Organize test commands
