+++ lib/commands/test.js
@@ -1,6 +1,6 @@
 const context = require('../cli-context');
 const stripes = require('@folio/stripes-core/webpack/stripes-node-api');
-const integrationTests = require('../integration-tests');
+const runIntegrationTests = require('../test/integration');
 const StripesPlatform = require('../platform/stripes-platform');
 const { applyOptions, serverOptions, stripesConfigOptions } = require('./common-options');
 const { cliAliases, cliResolve } = require('../webpack-common');
@@ -25,20 +25,28 @@ function testCommand(argv) {
     .then((stats) => {
       // TODO: Check for webpack errors before running tests
       console.log('Starting tests...');
-      integrationTests(argv); // TODO: then shutdown the server
+      runIntegrationTests(argv); // TODO: then shutdown the server
     });
 }
 
 module.exports = {
-  command: 'test',
-  describe: 'Run the current app module\'s tests (integration)',
+  command: 'test [config]',
+  describe: 'Run the current app module\'s tests',
   builder: (yargs) => {
     yargs
+      .positional('config', {
+        describe: 'File containing a Stripes tenant configuration',
+        type: 'string',
+      })
+      .option('run', {
+        describe: 'Name of the test script to run',
+        type: 'string',
+      })
       .option('show', {
         describe: 'Show UI and dev tools while running tests',
         type: 'boolean',
       })
-      .example('$0 --run=demo', 'Serve app and run it\'s demo.js integration tests');
+      .example('$0 test --run=demo', 'Serve app and run it\'s demo.js integration tests');
     return applyOptions(yargs, Object.assign({}, serverOptions, stripesConfigOptions));
   },
   handler: testCommand,
