+++ lib/commands/unit.js
@@ -1,27 +1,10 @@
 const context = require('../cli-context');
-const StripesConfigPlugin = require('@folio/stripes-core/webpack/stripes-config-plugin');
-const applyWebpackOverrides = require('@folio/stripes-core/webpack/apply-webpack-overrides');
 const unitTests = require('../test/unit');
 const StripesPlatform = require('../platform/stripes-platform');
 const { applyOptions, stripesConfigOptions } = require('./common-options');
 const { cliAliases, cliResolve } = require('../webpack-common');
-const webpack = require('webpack');
-const path = require('path');
+const getStripesWebpackConfig = require('../test/webpack-config');
 
-
-// TODO: Move this to stripes-core and expose as part of the Stripes Node API
-// Generates a webpack config for Stripes independent of build or serve
-function getStripesWebpackConfig(stripesConfig, options) {
-  // TODO: Switch between dev and prod files bases on env
-  let config = require('@folio/stripes-core/webpack.config.cli.dev'); // eslint-disable-line
-  config.plugins.push(new StripesConfigPlugin(stripesConfig));
-  config.plugins.push(new webpack.EnvironmentPlugin(['NODE_ENV']));
-  const platformModulePath = path.join(path.resolve(), 'node_modules'); // TODO: Verify
-  config.resolve.modules = ['node_modules', platformModulePath];
-  config.resolveLoader = { modules: ['node_modules', platformModulePath] };
-  config = applyWebpackOverrides(options.webpackOverrides, config);
-  return config;
-}
 
 function testCommand(argv) {
   if (context.type !== 'app') {
