+++ lib/test/cli-test-module.js
@@ -1,7 +1,8 @@
-// cli-test-module.js serves as a test loader called by mocha,
-// replicating behavior found in @folio/ui-testing/test-module.js
+// cli-test-module.js serves as a test loader called by mocha
+// to run an app's Nightmare-based integration tests.
+// This replicates behavior found in @folio/ui-testing/test-module.js
 //
-// This version modified in the following ways:
+// This version has been modified in the following ways:
 //  - Simplified to support limited options in use by stripes-cli
 //  - Supports cwd of the ui-app module under test
