+++ lib/test/integration.js
@@ -3,7 +3,7 @@ const path = require('path');
 
 // Creates a child process to run Mocha
 function runMocha(args) {
-  const mocha = path.join(__dirname, '../node_modules/.bin/mocha');
+  const mocha = path.join(__dirname, '../../node_modules/.bin/mocha');
   const cliTestModule = path.join(__dirname, 'cli-test-module.js');
   const cliArgs = [cliTestModule, '--exit'].concat(args);
 
@@ -26,7 +26,7 @@ function runMocha(args) {
 // Runs the specified integration tests
 // Recreates a subset of the options found ui-testing
 // However, '--run' only accepts a test name, rather than both the app and test name.
-module.exports = function integrationTests(options) {
+module.exports = function runIntegrationTests(options) {
   const testArgs = [
     '--run', options.run || 'test',
     '--url', `http://${options.host || 'localhost'}:${options.port || '3000'}`,
